Update PackagesAndPluginGuidelines.md
This documentation gives the impression that the example layout may use exact copies of some of the .ts files.
Line 103 and line 108 are both Action2.definition.ts.  Appending a "2" to Action implies a the action definition on 103 is somehow a copy of the one on 108.  This need not be true.  The "object1"s also imply this.   Only the "action1" and "action2" folders need have a unique name.  Removing the number elsewhere is more generic, cleaner to read, and easier to understand.